Refactoring mainloop.
