+++ src/tools/mainloop/os/linux/handler.cc
@@ -18,6 +18,7 @@
  */
 
  #include <config.h>
+ #include <private/linux/mainloop.h>
  #include <udjat/tools/handler.h>
  #include <poll.h>
  #include <system_error>
